Adds better handling for image uploading
Adds handling for status 413 (request payload too large) so that
information is displayed to the user if the server rejects the
image on the basis of file size.

Uploaded image now shows immediately (using base64 data from image),
instead of waiting for the image upload to complete before
displaying. Doing this required handling for concurrent image
uploads and handling a user clicking 'clear image' before
the image upload had completed. These cases are handled with
an uploadId that tracks if the upload is current or not.